pyproject.toml
[tool.pylint.FORMAT]
max-line-length = 119

[tool.poetry]
name = "moneyed-fx"
version = "0.0.1"
description = "FX handler for Jenfi's use of py-moneyed"
authors = ["Justin Louie <224840+nitsujri@users.noreply.github.com>"]
readme = "README.md"
packages = [{include = "moneyed_fx"}]

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.6"
py-moneyed = "^3.0"
psycopg2 = "^2.9.9"
django-model-utils = "^4.3.1"


[tool.poetry.group.dev.dependencies]
pre-commit = "^3.5.0"
pytest = "^7.4.3"
flake8 = "^6.1.0"
flake8-isort = "^6.1.0"
black = "^23.10.1"
pylint-django = "^2.5.5"
factory-boy = "^3.3.0"
pytest-django = "^4.5.2"
pytest-randomly = "^3.15.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
